Ricardo: fix bugs
